p29.22 refactor: thread ids.ImportBatchID + ids.ImportRowID everywhere
Typed-ID capstone: ImportBatchID and ImportRowID (coupled -- a row carries its
batch FK) are now distinct types end-to-end, following the ReconciliationID
template. Purely types, no behavior change.

Threaded:
- sqlc: ImportBatch.ID, ImportRow.ID + ImportRow.BatchID (the coupled FK);
  GetImportBatch/ImportRowsByBatch/GetImportRow params + GetImportRowRow.ID/
  .BatchID; InsertImportBatch/InsertImportRow returns; InsertImportRowParams.
  BatchID; MarkImportRowDiscarded param; MarkImportRowPostedParams.ID. No
  in-body wraps needed (database/sql scans/args the typed ids via reflection).
- store: CreateImportBatch/InsertImportRow returns + newID; StageImportRows
  batchID; StagedRow.ID; ImportBatchRow.ID; ImportRowsForBatch(Flagged)/
  GetImportBatch batch param + store.ImportBatch.ID; store.ImportRow.ID/.BatchID;
  GetImportRow/PostImportRow/DiscardImportRow rowID param.
- web: parse to typed at the parseID seam (ids.ImportBatchID/ImportRowID(...));
  render-only model fields (importQueueModel.BatchID, txnEditorModel.Import*ID,
  importResultModel.BatchID, importQueueRow.ID) stay int64 -- convert with
  int64(...) at assignment + every strconv.FormatInt URL build.
- synth: DemoIDs.ImportBatch typed. tests: stageOnePending return, reviewEnv
  {rowID,batchID}, raw-SQL scan target typed; FormatInt call sites wrapped.

Reserved for their own commits (left bare this commit): ImportBatch/Row.AccountID
+ SubsidiaryID (Fund/Sub batches unrelated -- these are AccountID/SubsidiaryID),
ProfileID + all MappingProfile* (commit 2), UploadedBy (UserID, commit 3),
PostedTransactionID (NullInt64, TransactionID final batch), and PostImportRow's
returned txnID (TransactionID).

+++ internal/web/imports.go
@@ -667,7 +667,7 @@ func (s *server) importConfirm(w http.ResponseWriter, r *http.Request) {
 	// The result is swapped into #import-workspace via htmx innerHTML (replacing the
 	// preview), so render JUST the fragment (no shell).
 	s.render(w, r, http.StatusOK, "import-result",
-		s.buildImportResult(r, batchID, filename, acct.DefaultCurrency, exp, staged))
+		s.buildImportResult(r, int64(batchID), filename, acct.DefaultCurrency, exp, staged))
 }
 
 // buildImportResult formats the staged-row result for the confirm page.
